Fix a bug with POST and PUT request params and bodies
Data should be passed as parameters in the body of the request instead
of converting it into query string params. Query string will work for
smaller amounts of data, but errors arise when the query string becomes
too long. For example this is often the case when requesting deltas
using a long cursor string.